pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signalflow"
version = "1.0.0"
description = "Real-Time Stock Market Prediction System using LSTM Neural Networks, Apache Kafka, and MLflow"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Developer"}
]
keywords = ["stock-prediction", "lstm", "kafka", "mlflow", "pytorch", "machine-learning", "trading", "finance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "ta>=0.10.2",
    "kafka-python>=2.0.2",
    "mlflow>=2.8.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "yfinance>=0.2.30",
    "structlog>=23.2.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "tqdm>=4.66.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "aiokafka>=0.9.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.5",
    "mypy>=1.7.0",
]

[project.scripts]
replay-service = "src.replay_service.main:main"
feature-service = "src.feature_service.main:main"
train-model = "src.training.train:main"
inference-service = "src.inference_service.main:main"
download-data = "src.data.download:main"
dashboard = "src.dashboard.server:main"
live-service = "src.live_service.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "configs*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

